I have updated my code again
Now my code is able to detect what java file you are working in, this is done with the method called getFileName(String args[]). Before I had the java file to be hard coded to one java file but now you can make as many files you need and it can tell you what file you are working on + it can extract everything in the file you are working in.